refactor: split server, client and tui by type
