+++ cmd/plan.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/marcbran/versource/internal"
-	"github.com/marcbran/versource/internal/http"
+	"github.com/marcbran/versource/internal/http/client"
 	"github.com/spf13/cobra"
 )
 
@@ -35,7 +35,7 @@ var planCmd = &cobra.Command{
 			return err
 		}
 
-		client := http.NewClient(config)
+		client := client.NewClient(config)
 
 		req := internal.CreatePlanRequest{
 			ComponentID: componentID,
