+++ internal/http/client/client.go
@@ -1,0 +1,25 @@
+package client
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/marcbran/versource/internal"
+)
+
+type Client struct {
+	baseURL string
+	client  *http.Client
+}
+
+func NewClient(config *internal.Config) *Client {
+	baseURL := fmt.Sprintf("%s://%s:%s", config.HTTP.Scheme, config.HTTP.Hostname, config.HTTP.Port)
+	if config.HTTP.Hostname == "" {
+		baseURL = fmt.Sprintf("%s://localhost:%s", config.HTTP.Scheme, config.HTTP.Port)
+	}
+
+	return &Client{
+		baseURL: baseURL,
+		client:  &http.Client{},
+	}
+}
