+++ internal/http/server/changeset.go
@@ -1,0 +1,57 @@
+package server
+
+import (
+	"encoding/json"
+	"fmt"
+	"net/http"
+
+	"github.com/go-chi/chi/v5"
+	"github.com/marcbran/versource/internal"
+)
+
+func (s *Server) handleListChangesets(w http.ResponseWriter, r *http.Request) {
+	resp, err := s.listChangesets.Exec(r.Context(), internal.ListChangesetsRequest{})
+	if err != nil {
+		returnError(w, err)
+		return
+	}
+
+	returnSuccess(w, resp)
+}
+
+func (s *Server) handleCreateChangeset(w http.ResponseWriter, r *http.Request) {
+	var req internal.CreateChangesetRequest
+	err := json.NewDecoder(r.Body).Decode(&req)
+	if err != nil {
+		returnBadRequest(w, fmt.Errorf("invalid request body"))
+		return
+	}
+
+	resp, err := s.createChangeset.Exec(r.Context(), req)
+	if err != nil {
+		returnError(w, err)
+		return
+	}
+
+	returnCreated(w, resp)
+}
+
+func (s *Server) handleMergeChangeset(w http.ResponseWriter, r *http.Request) {
+	changesetName := chi.URLParam(r, "changesetName")
+	if changesetName == "" {
+		returnBadRequest(w, fmt.Errorf("changeset name is required"))
+		return
+	}
+
+	req := internal.MergeChangesetRequest{
+		ChangesetName: changesetName,
+	}
+
+	resp, err := s.mergeChangeset.Exec(r.Context(), req)
+	if err != nil {
+		returnError(w, err)
+		return
+	}
+
+	returnSuccess(w, resp)
+}
