+++ internal/tui/app.go
@@ -1,9 +1,7 @@
 package tui
 
 import (
-	"context"
 	"fmt"
-	"strconv"
 	"strings"
 
 	"github.com/charmbracelet/bubbles/table"
@@ -11,11 +9,11 @@ import (
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
 	"github.com/marcbran/versource/internal"
-	"github.com/marcbran/versource/internal/http"
+	"github.com/marcbran/versource/internal/http/client"
 )
 
 type App struct {
-	client      *http.Client
+	client      *client.Client
 	router      *Router
 	currentView string
 	viewHistory []string
@@ -43,7 +41,7 @@ type Rect struct {
 	Height int
 }
 
-func NewApp(client *http.Client) *App {
+func NewApp(client *client.Client) *App {
 	ti := textinput.New()
 	ti.CharLimit = 100
 
@@ -179,148 +177,6 @@ func getTable(data any) ([]table.Column, []table.Row, []string) {
 	default:
 		return []table.Column{}, []table.Row{}, []string{}
 	}
-}
-
-func getModulesTable(modules []internal.Module) ([]table.Column, []table.Row, []string) {
-	columns := []table.Column{
-		{Title: "ID", Width: 1},
-		{Title: "Source", Width: 9},
-	}
-
-	var rows []table.Row
-	var ids []string
-	for _, module := range modules {
-		rows = append(rows, table.Row{
-			strconv.FormatUint(uint64(module.ID), 10),
-			module.Source,
-		})
-		ids = append(ids, strconv.FormatUint(uint64(module.ID), 10))
-	}
-
-	return columns, rows, ids
-}
-
-func getModuleVersionsTable(moduleVersions []internal.ModuleVersion) ([]table.Column, []table.Row, []string) {
-	columns := []table.Column{
-		{Title: "ID", Width: 1},
-		{Title: "Module", Width: 7},
-		{Title: "Version", Width: 2},
-	}
-
-	var rows []table.Row
-	var ids []string
-	for _, moduleVersion := range moduleVersions {
-		source := ""
-		if moduleVersion.Module.Source != "" {
-			source = moduleVersion.Module.Source
-		}
-		rows = append(rows, table.Row{
-			strconv.FormatUint(uint64(moduleVersion.ID), 10),
-			source,
-			moduleVersion.Version,
-		})
-		ids = append(ids, strconv.FormatUint(uint64(moduleVersion.ID), 10))
-	}
-
-	return columns, rows, ids
-}
-
-func getChangesetsTable(changesets []internal.Changeset) ([]table.Column, []table.Row, []string) {
-	columns := []table.Column{
-		{Title: "ID", Width: 1},
-		{Title: "Name", Width: 7},
-		{Title: "State", Width: 2},
-		{Title: "Review", Width: 2},
-	}
-
-	var rows []table.Row
-	var ids []string
-	for _, changeset := range changesets {
-		rows = append(rows, table.Row{
-			strconv.FormatUint(uint64(changeset.ID), 10),
-			changeset.Name,
-			string(changeset.State),
-			string(changeset.ReviewState),
-		})
-		ids = append(ids, strconv.FormatUint(uint64(changeset.ID), 10))
-	}
-
-	return columns, rows, ids
-}
-
-func getComponentsTable(components []internal.Component) ([]table.Column, []table.Row, []string) {
-	columns := []table.Column{
-		{Title: "ID", Width: 1},
-		{Title: "Module", Width: 7},
-		{Title: "Version", Width: 2},
-	}
-
-	var rows []table.Row
-	var ids []string
-	for _, component := range components {
-		source := ""
-		version := ""
-		if component.ModuleVersion.Module.Source != "" {
-			source = component.ModuleVersion.Module.Source
-		}
-		if component.ModuleVersion.Version != "" {
-			version = component.ModuleVersion.Version
-		}
-		rows = append(rows, table.Row{
-			strconv.FormatUint(uint64(component.ID), 10),
-			source,
-			version,
-		})
-		ids = append(ids, strconv.FormatUint(uint64(component.ID), 10))
-	}
-
-	return columns, rows, ids
-}
-
-func getPlansTable(plans []internal.Plan) ([]table.Column, []table.Row, []string) {
-	columns := []table.Column{
-		{Title: "ID", Width: 1},
-		{Title: "Component", Width: 1},
-		{Title: "Changeset", Width: 6},
-		{Title: "State", Width: 2},
-	}
-
-	var rows []table.Row
-	var ids []string
-	for _, plan := range plans {
-		rows = append(rows, table.Row{
-			strconv.FormatUint(uint64(plan.ID), 10),
-			strconv.FormatUint(uint64(plan.ComponentID), 10),
-			plan.Changeset.Name,
-			plan.State,
-		})
-		ids = append(ids, strconv.FormatUint(uint64(plan.ID), 10))
-	}
-
-	return columns, rows, ids
-}
-
-func getAppliesTable(applies []internal.Apply) ([]table.Column, []table.Row, []string) {
-	columns := []table.Column{
-		{Title: "ID", Width: 1},
-		{Title: "Plan", Width: 1},
-		{Title: "Changeset", Width: 6},
-		{Title: "State", Width: 2},
-	}
-
-	var rows []table.Row
-	var ids []string
-	for _, apply := range applies {
-		rows = append(rows, table.Row{
-			strconv.FormatUint(uint64(apply.ID), 10),
-			strconv.FormatUint(uint64(apply.PlanID), 10),
-			apply.Changeset.Name,
-			apply.State,
-		})
-		ids = append(ids, strconv.FormatUint(uint64(apply.ID), 10))
-	}
-
-	return columns, rows, ids
 }
 
 func createTable(columns []table.Column, rows []table.Row, size Rect, showInput bool) table.Model {
@@ -489,305 +345,4 @@ func titledBox(title, content string) string {
 		Render(content)
 
 	return lipgloss.JoinVertical(lipgloss.Left, top, body)
-}
-
-type ModulesPage struct {
-	app *App
-}
-
-func (p *ModulesPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		resp, err := p.app.client.ListModules(ctx)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-		return dataLoadedMsg{view: "modules", data: resp.Modules}
-	}
-}
-
-func (p *ModulesPage) Links(params map[string]string) map[string]string {
-	return map[string]string{}
-}
-
-type ModulePage struct {
-	app *App
-}
-
-func (p *ModulePage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		return dataLoadedMsg{view: fmt.Sprintf("modules/%s", params["moduleID"]), data: nil}
-	}
-}
-
-func (p *ModulePage) Links(params map[string]string) map[string]string {
-	return map[string]string{
-		"enter": fmt.Sprintf("modules/%s/moduleversions", params["moduleID"]),
-		"c":     fmt.Sprintf("components?module-id=%s", params["moduleID"]),
-	}
-}
-
-type ModuleVersionsPage struct {
-	app *App
-}
-
-func (p *ModuleVersionsPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		resp, err := p.app.client.ListModuleVersions(ctx)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-		return dataLoadedMsg{view: "moduleversions", data: resp.ModuleVersions}
-	}
-}
-
-func (p *ModuleVersionsPage) Links(params map[string]string) map[string]string {
-	return map[string]string{
-		"c": fmt.Sprintf("components?module-version-id=%s", params["moduleVersionID"]),
-	}
-}
-
-type ModuleVersionsForModulePage struct {
-	app *App
-}
-
-func (p *ModuleVersionsForModulePage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		moduleID, exists := params["moduleID"]
-		if !exists {
-			return errorMsg{err: fmt.Errorf("moduleID parameter required")}
-		}
-
-		moduleIDUint, err := strconv.ParseUint(moduleID, 10, 32)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-		resp, err := p.app.client.ListModuleVersionsForModule(ctx, uint(moduleIDUint))
-		if err != nil {
-			return errorMsg{err: err}
-		}
-		return dataLoadedMsg{view: fmt.Sprintf("modules/%s/moduleversions", moduleID), data: resp.ModuleVersions}
-	}
-}
-
-func (p *ModuleVersionsForModulePage) Links(params map[string]string) map[string]string {
-	return map[string]string{
-		"m": "modules",
-	}
-}
-
-type ChangesetsPage struct {
-	app *App
-}
-
-func (p *ChangesetsPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		resp, err := p.app.client.ListChangesets(ctx)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-		return dataLoadedMsg{view: "changesets", data: resp.Changesets}
-	}
-}
-
-func (p *ChangesetsPage) Links(params map[string]string) map[string]string {
-	return map[string]string{}
-}
-
-type ChangesetPage struct {
-	app *App
-}
-
-func (p *ChangesetPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		return dataLoadedMsg{view: fmt.Sprintf("changesets/%s", params["changesetName"]), data: nil}
-	}
-}
-
-func (p *ChangesetPage) Links(params map[string]string) map[string]string {
-	changesetName := params["changesetName"]
-	return map[string]string{
-		"enter": fmt.Sprintf("changesets/%s/components", changesetName),
-		"c":     fmt.Sprintf("changesets/%s/components", changesetName),
-		"p":     fmt.Sprintf("changesets/%s/plans", changesetName),
-		"a":     fmt.Sprintf("changesets/%s/applies", changesetName),
-	}
-}
-
-type ComponentsPage struct {
-	app *App
-}
-
-func (p *ComponentsPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-
-		req := internal.ListComponentsRequest{}
-
-		if moduleIDStr, ok := params["module-id"]; ok {
-			if moduleID, err := strconv.ParseUint(moduleIDStr, 10, 32); err == nil {
-				moduleIDUint := uint(moduleID)
-				req.ModuleID = &moduleIDUint
-			}
-		}
-
-		if moduleVersionIDStr, ok := params["module-version-id"]; ok {
-			if moduleVersionID, err := strconv.ParseUint(moduleVersionIDStr, 10, 32); err == nil {
-				moduleVersionIDUint := uint(moduleVersionID)
-				req.ModuleVersionID = &moduleVersionIDUint
-			}
-		}
-
-		resp, err := p.app.client.ListComponents(ctx, req)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-
-		view := "components"
-		if len(params) > 0 {
-			queryParts := make([]string, 0)
-			if moduleIDStr, ok := params["module-id"]; ok {
-				queryParts = append(queryParts, fmt.Sprintf("module-id=%s", moduleIDStr))
-			}
-			if moduleVersionIDStr, ok := params["module-version-id"]; ok {
-				queryParts = append(queryParts, fmt.Sprintf("module-version-id=%s", moduleVersionIDStr))
-			}
-			if len(queryParts) > 0 {
-				view = fmt.Sprintf("components?%s", strings.Join(queryParts, "&"))
-			}
-		}
-
-		return dataLoadedMsg{view: view, data: resp.Components}
-	}
-}
-
-func (p *ComponentsPage) Links(params map[string]string) map[string]string {
-	return map[string]string{
-		"m": "modules",
-		"v": "moduleversions",
-	}
-}
-
-type PlansPage struct {
-	app *App
-}
-
-func (p *PlansPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		resp, err := p.app.client.ListPlans(ctx)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-		return dataLoadedMsg{view: "plans", data: resp.Plans}
-	}
-}
-
-func (p *PlansPage) Links(params map[string]string) map[string]string {
-	return map[string]string{}
-}
-
-type AppliesPage struct {
-	app *App
-}
-
-func (p *AppliesPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		resp, err := p.app.client.ListApplies(ctx)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-		return dataLoadedMsg{view: "applies", data: resp.Applies}
-	}
-}
-
-func (p *AppliesPage) Links(params map[string]string) map[string]string {
-	return map[string]string{}
-}
-
-type ChangesetComponentsPage struct {
-	app *App
-}
-
-func (p *ChangesetComponentsPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		changesetName := params["changesetName"]
-
-		req := internal.ListComponentsRequest{}
-
-		resp, err := p.app.client.ListComponents(ctx, req)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-
-		view := fmt.Sprintf("changesets/%s/components", changesetName)
-		return dataLoadedMsg{view: view, data: resp.Components}
-	}
-}
-
-func (p *ChangesetComponentsPage) Links(params map[string]string) map[string]string {
-	changesetName := params["changesetName"]
-	return map[string]string{
-		"p": fmt.Sprintf("changesets/%s/plans", changesetName),
-		"a": fmt.Sprintf("changesets/%s/applies", changesetName),
-	}
-}
-
-type ChangesetPlansPage struct {
-	app *App
-}
-
-func (p *ChangesetPlansPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		changesetName := params["changesetName"]
-
-		resp, err := p.app.client.ListPlans(ctx)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-
-		view := fmt.Sprintf("changesets/%s/plans", changesetName)
-		return dataLoadedMsg{view: view, data: resp.Plans}
-	}
-}
-
-func (p *ChangesetPlansPage) Links(params map[string]string) map[string]string {
-	changesetName := params["changesetName"]
-	return map[string]string{
-		"c": fmt.Sprintf("changesets/%s/components", changesetName),
-		"a": fmt.Sprintf("changesets/%s/applies", changesetName),
-	}
-}
-
-type ChangesetAppliesPage struct {
-	app *App
-}
-
-func (p *ChangesetAppliesPage) Open(params map[string]string) tea.Cmd {
-	return func() tea.Msg {
-		ctx := context.Background()
-		changesetName := params["changesetName"]
-
-		resp, err := p.app.client.ListApplies(ctx)
-		if err != nil {
-			return errorMsg{err: err}
-		}
-
-		view := fmt.Sprintf("changesets/%s/applies", changesetName)
-		return dataLoadedMsg{view: view, data: resp.Applies}
-	}
-}
-
-func (p *ChangesetAppliesPage) Links(params map[string]string) map[string]string {
-	changesetName := params["changesetName"]
-	return map[string]string{
-		"c": fmt.Sprintf("changesets/%s/components", changesetName),
-		"p": fmt.Sprintf("changesets/%s/plans", changesetName),
-	}
 }
